Rename `data-models` to `models/data` (#532)
