+++ v3/src/components/app.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react"
 import React from "react"
-import { gDataBroker } from "../data-model/data-broker"
+import { gDataBroker } from "../models/data/data-broker"
 import { convertParsedCsvToDataSet } from "../utilities/csv-import"
 import { prf } from "../utilities/profiler"
 import { setUrlParams } from "../utilities/url-params"
