+++ v3/src/components/app.tsx
@@ -4,8 +4,8 @@ import { CodapDndContext } from "./codap-dnd-context"
 import { ToolShelf } from "./tool-shelf/tool-shelf"
 import {Container} from "./container"
 import {DataSummary} from "./data-summary"
-import {gDataBroker} from "../data-model/data-broker"
-import {DataSet, IDataSet, toCanonical} from "../data-model/data-set"
+import {gDataBroker} from "../models/data/data-broker"
+import {DataSet, IDataSet, toCanonical} from "../models/data/data-set"
 import { GraphComponent } from "./graph/components/graph-component"
 import {Text} from "./text"
 import {useDropHandler} from "../hooks/use-drop-handler"
