+++ v3/src/components/case-table/use-rows.ts
@@ -2,11 +2,11 @@ import { format } from "d3"
 import { reaction } from "mobx"
 import { onAction } from "mobx-state-tree"
 import { useCallback, useEffect, useMemo, useRef, useState } from "react"
-import { kDefaultFormatStr } from "../../data-model/attribute"
-import { ICase, IDataSet } from "../../data-model/data-set"
+import { kDefaultFormatStr } from "../../models/data/attribute"
+import { ICase, IDataSet } from "../../models/data/data-set"
 import {
   AddCasesAction, isRemoveCasesAction, RemoveCasesAction, SetCaseValuesAction
-} from "../../data-model/data-set-actions"
+} from "../../models/data/data-set-actions"
 import { prf } from "../../utilities/profiler"
 import { appState } from "../app-state"
 import { TRow, TRowsChangeData } from "./case-table-types"
