+++ v3/src/components/case-table/use-selected-rows.ts
@@ -2,8 +2,8 @@ import { reaction } from "mobx"
 import { onAction } from "mobx-state-tree"
 import { useCallback, useEffect, useRef, useState } from "react"
 import { appState } from "../app-state"
-import { IDataSet } from "../../data-model/data-set"
-import { isPartialSelectionAction, isSelectionAction } from "../../data-model/data-set-actions"
+import { IDataSet } from "../../models/data/data-set"
+import { isPartialSelectionAction, isSelectionAction } from "../../models/data/data-set-actions"
 import { prf } from "../../utilities/profiler"
 import { isKeyDown } from "../../hooks/use-key-states"
 import { TRow } from "./case-table-types"
