+++ v3/src/components/data-summary.tsx
@@ -2,8 +2,8 @@ import { Button, Select } from '@chakra-ui/react'
 import { Active, DragOverlay, useDndContext, useDroppable } from "@dnd-kit/core"
 import { observer } from "mobx-react-lite"
 import React, { useState } from "react"
-import { IAttribute } from "../data-model/attribute"
-import { DataBroker } from "../data-model/data-broker"
+import { IAttribute } from "../models/data/attribute"
+import { DataBroker } from "../models/data/data-broker"
 import { getDragAttributeId, IDropData, IUseDraggableAttribute, useDraggableAttribute } from '../hooks/use-drag-drop'
 import { useV2DocumentContext } from '../hooks/use-v2-document-context'
 import { prf } from "../utilities/profiler"
