+++ v3/src/components/graph/components/casedots.tsx
@@ -2,8 +2,8 @@ import {randomUniform, select} from "d3"
 import { onAction } from "mobx-state-tree"
 import React, {memo, useCallback, useEffect, useRef, useState} from "react"
 import {pointRadiusSelectionAddend, transitionDuration} from "../graphing-types"
-import { ICase } from "../../../data-model/data-set"
-import { isAddCasesAction } from "../../../data-model/data-set-actions"
+import { ICase } from "../../../models/data/data-set"
+import { isAddCasesAction } from "../../../models/data/data-set-actions"
 import {useDragHandlers, usePlotResponders} from "../hooks/graph-hooks"
 import {useDataConfigurationContext} from "../hooks/use-data-configuration-context"
 import {useDataSetContext} from "../../../hooks/use-data-set-context"
