+++ v3/src/components/graph/components/graph-component.tsx
@@ -3,7 +3,7 @@ import {observer} from "mobx-react-lite"
 import {getSnapshot} from "mobx-state-tree"
 import React, {useEffect, useMemo, useRef} from "react"
 import {useResizeDetector} from "react-resize-detector"
-import {DataBroker} from "../../../data-model/data-broker"
+import {DataBroker} from "../../../models/data/data-broker"
 import {DataSetContext} from "../../../hooks/use-data-set-context"
 import {InstanceIdContext, useNextInstanceId} from "../../../hooks/use-instance-id-context"
 import {EmptyAxisModel} from "../models/axis-model"
