+++ v3/src/components/graph/utilities/graph-utils.test.ts
@@ -1,5 +1,5 @@
 import {equationString, getScreenCoord, lineToAxisIntercepts, ptInRect, valueLabelString} from "./graph-utils"
-import {DataSet, toCanonical} from "../../../data-model/data-set"
+import {DataSet, toCanonical} from "../../../models/data/data-set"
 import {scaleLinear} from "d3"
 
 describe("equationString", () => {
